+++ lizard_area/admin.py
@@ -11,7 +11,7 @@ from lizard_area.models import SynchronizationHistory
 from lizard_area.models import DataAdministrator
 from lizard_area.models import Communique
 from lizard_area.models import Area
-
+from lizard_area.models import AreaLink
 
 class GeoObjectInline(admin.TabularInline):
     model = GeoObject
@@ -27,6 +27,9 @@ class AreaAdmin(admin.ModelAdmin):
     list_filter = ('data_set', 'area_class', 'is_active', )
     list_display = ('name', 'ident', 'id', 'area_class', 'is_active', )
 
+class AreaLinkAdmin(admin.ModelAdmin):
+    list_filter = ('area_a', 'area_b', )
+    list_display = ('area_a', 'area_b', )
 
 # admin.site.register(GeoObject)
 admin.site.register(GeoObjectGroup, GeoObjectGroupAdmin)
@@ -37,5 +40,6 @@ admin.site.register(MapnikXMLStyleSheet)
 admin.site.register(DataAdministrator)
 admin.site.register(Communique)
 admin.site.register(Area, AreaAdmin)
+admin.site.register(AreaLink, AreaLinkAdmin)
 admin.site.register(AreaWFSConfiguration)
 admin.site.register(SynchronizationHistory)